requirements.txt
psycopg==3.2.9
psycopg-binary==3.2.9
python-dotenv==1.1.1
